requirements.txt
requests
psycopg2-binary
python-dotenv
apscheduler
beautifulsoup4
fastapi
uvicorn
sqlalchemy
